fix precision issue

+++ social_dilemmas/envs/cleanup.py
@@ -84,7 +84,7 @@ class CleanupEnv(MapEnv):
     @property
     def action_space(self):
         # return DiscreteWithDType(9, dtype=np.uint8)
-        return DiscreteWithDType(9, dtype=np.int64)
+        return DiscreteWithDType(9, dtype=np.int32)
 
     def custom_reset(self):
         """Initialize the walls and the waste"""
